Initial structure and test layout
