+++ server/lib/src/constants/uuids.rs
@@ -83,6 +83,7 @@ pub const UUID_IDM_MESSAGE_SENDERS: Uuid = uuid!("00000000-0000-0000-0000-000000
 pub const UUID_IDM_OAUTH2_CLIENT_ADMINS: Uuid = uuid!("00000000-0000-0000-0000-000000000056");
 pub const UUID_IDM_OAUTH2_ACCOUNT_ADMINS: Uuid = uuid!("00000000-0000-0000-0000-000000000057");
 pub const UUID_HMAC_NAME_FEATURE: Uuid = uuid!("00000000-0000-0000-0000-000000000058");
+pub const UUID_ACCOUNT_SIGNUP_FEATURE: Uuid = uuid!("00000000-0000-0000-0000-000000000059");
 
 //
 pub const UUID_IDM_HIGH_PRIVILEGE: Uuid = uuid!("00000000-0000-0000-0000-000000001000");
